# RealEstateAgency.WebApplication/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Client API",
    "description": "API для работы с клиентами",
    "version": "v1"
  },
  "paths": {
    "/api/Analytics/clients-by-realestate-type": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Вывести сведения о всех клиентах, ищущих недвижимость заданного типа, упорядоченные по ФИО.",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Значения для параметра `type` могут быть следующими:\r\n<list type=\"bullet\"><item><term>Residential</term></item><item><term>Uninhabitable</term></item></list>",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Analytics/sellers-by-period": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Вывести всех продавцов, оставивших заявки за заданный период.",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "description": "Начальная дата периода.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "Конечная дата периода.",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Analytics/matching-sellers-for-buyer/{buyerOrderId}": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Вывести сведения о продавцах и объектах недвижимости, соответствующих определенной заявке покупателя.",
        "parameters": [
          {
            "name": "buyerOrderId",
            "in": "path",
            "description": "Идентификатор заявки покупателя.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SellerRealEstateDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerRealEstateDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerRealEstateDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Analytics/order-count-by-type": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Вывести информацию о количестве заявок по каждому типу недвижимости.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RealEstateOrderCountDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RealEstateOrderCountDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RealEstateOrderCountDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Analytics/top-purchasers": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Вывести топ 5 клиентов по количеству заявок на покупку.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientOrderCountDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientOrderCountDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientOrderCountDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Analytics/top-sellers": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Вывести топ 5 клиентов по количеству заявок на продажу.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientOrderCountDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientOrderCountDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientOrderCountDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Analytics/min-price-orders": {
      "get": {
        "tags": [
          "Analytics"
        ],
        "summary": "Вывести информацию о клиентах, открывших заявки с минимальной стоимостью.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientOrderPriceDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientOrderPriceDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientOrderPriceDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ClientDto": {
        "type": "object",
        "properties": {
          "firstAndLastName": {
            "type": "string",
            "nullable": true
          },
          "pasport": {
            "type": "string",
            "nullable": true
          },
          "numberPhone": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ClientOrderCountDto": {
        "type": "object",
        "properties": {
          "client": {
            "$ref": "#/components/schemas/ClientDto"
          },
          "orderCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ClientOrderPriceDto": {
        "type": "object",
        "properties": {
          "client": {
            "$ref": "#/components/schemas/ClientDto"
          },
          "orderPrice": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "RealEstateDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "nullable": true
          },
          "square": {
            "type": "number",
            "format": "double"
          },
          "numberOfRooms": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RealEstateOrderCountDto": {
        "type": "object",
        "properties": {
          "realEstateType": {
            "type": "string",
            "nullable": true
          },
          "orderCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SellerRealEstateDto": {
        "type": "object",
        "properties": {
          "realEstate": {
            "$ref": "#/components/schemas/RealEstateDto"
          },
          "sellers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}